use f32 as test type

+++ hpt-bench/benches/benchmarks/binary/matmul_f32.rs
@@ -27,8 +27,8 @@ fn matmul_f32_benchmark(c: &mut Criterion<crate::benchmarks::Timer>) {
     for n in ns {
         // let a = black_box(TchTensor::randn([n, n], (Kind::Float, Device::Cpu)));
         // let c = black_box(TchTensor::randn([n, n], (Kind::Float, Device::Cpu)));
-        let a2 = black_box(Tensor::<f16>::randn([n, n]).unwrap());
-        let c2 = black_box(Tensor::<f16>::randn([n, n]).unwrap());
+        let a2 = black_box(Tensor::<f32>::randn([n, n]).unwrap());
+        let c2 = black_box(Tensor::<f32>::randn([n, n]).unwrap());
         // let a3 = black_box(
         //     CandleTensor::randn(
         //         0.0f32,
